% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{compactify_chain_multi_labels}
\alias{compactify_chain_multi_labels}
\title{Compactify mulilayer labels}
\usage{
compactify_chain_multi_labels(zh_list)
}
\arguments{
\item{zh_list}{a list of lists of labels}
}
\description{
The function takes as input a list of lists of labels and collapses it to a
matrix. The outer list iterates over chain iterations, and the inner list
iterates over the layers. The output is a matrix of dimension
number_of_node_layer_pairs x number_of_iterations
}
